+++ App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import StartMenu from './components/startMenu';
 
-import MenuButton from './components/Menu_Bottom';
+import MenuButton from './components/MenuBottom';
 import { View } from 'react-native';
 
 export default function App() {
